pyproject.toml
[tool.poetry]
name = "mini-app-tg-gpt"
version = "0.1.0"
description = ""
authors = ["l1ghtny <ankoshenkov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
openai = ">=1.109.1"
python-dotenv = ">=1.0.1"
sqlmodel = ">=0.0.25"
SQLAlchemy = ">=2.0.30"
python-telegram-bot = ">=21.4"
requests = ">=2.31.0"
typing_extensions = ">=4.11.0"
fastapi = {extras = ["standard"], version = "^0.118.0"}
pydantic = ">=2.5.2"
alembic = "^1.17.0"
redis = "^7.0.1"
sse-starlette = "^3.0.2"
aioboto3 = "^15.5.0"
pillow = "^12.0.0"
pillow-heif = "^1.1.1"
fastapi-swagger-dark = "^0.0.8"
python-jose = "^3.5.0"
passlib = "^1.7.4"
sentry-sdk = {extras = ["fastapi"], version = "^2.47.0"}
aiogram = "^3.23.0"
uuid6 = "^2025.0.1"
asyncpg = "^0.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
